Storage Class Updates
Updated the Storage so that it should hopefully run through the whole
excel sheet no matter how many items are in it. It should also hopefully
be able to pull multiple images for a ShoppingItem from the excel sheet
no matter how many it has.

Note: As I'm writing this the multiple images probably doesn't work but ill work on that right now.