all the base things needed to add multiple images
multiple images isn't implemented fully yet but the single images do use
a queue and are added to a queue when creating their ShoppingItem. I
just wanted to push so that once i start trying to actually implement
the multiple images I have a version to revert to in case things go
horribly wrong.